+++ include/huira_impl/handles/camera_handle.ipp
@@ -2,14 +2,14 @@
 
 namespace huira {
     template <IsSpectral TSpectral>
-    void CameraModelHandle<TSpectral>::set_focal_length(double focal_length) const {
+    void CameraModelHandle<TSpectral>::set_focal_length(float focal_length) const {
         this->get()->set_focal_length(focal_length);
     }
 
     template <IsSpectral TSpectral>
-    void CameraModelHandle<TSpectral>::set_fstop(double fstop) const
+    void CameraModelHandle<TSpectral>::set_fstop(float fstop) const
     {
-        this->get()->set_fstop(static_cast<float>(fstop));
+        this->get()->set_fstop(fstop);
     }
 
     template <IsSpectral TSpectral>
